Exposed CreateSelfLinkFrom as a public helper
method.
Why is this a static method? It looks like Feature Envy.
The reason is that while it could technically have been placed on the
UuidIri struct, it's not a generally applicable method for all UuidIri
instances. Rather, it's closely related to how AtomEventStream<T> handles
and expects the relationship to be between IDs and self links.
This is also the reason this is not an extension method, because unless I
would have placed this extension method in a new (very small) namespace,
the CreateSelfLink method would have appeared on all UuidIri variables, so
that would have had the same effect as putting the method on UuidIri.