Kept existing entries in feed,
instead of replacing a single entry in the feed. This will allow the feed
to grow unbounded, so a paging mechanism is required next. Currently,
entries are kept both as entries within the feed, and as separate,
addressable items, but when the feed paging algorithm is correctly
implemented, items will no longer be saved individually.