Changed file name algorithm for file store
to always use the .xml extension. This is because it annoys me that I
can't easily open the files on my file system, because until now, they've
had no extension.

While this is a breaking change, I don't believe that we have any
production systems running on files. Existing files can still be used if
they are renamed to get the .xml file extension.