Added .xml extension to blobs,
partly because it's nicer if you want to download the blobs to a
development machine, or use other file-based tools for managing the blobs;
and partly to stay consistent with the similarly suggested changed in the
file-based storage for AtomEventStore.

Please not that this is a breaking change. However, we should be able to
migrate our current blobs by appending their names with the .xml
extension.