Incremented minor version,
in order to indicate that the addition of FifoEvents<T> is a new feature.
Also, there are breaking changes in AtomEventsOnAzure and
AtomEventsInFiles, because they now use extensions as part of their
internal feed addresses. However, since this is major version 0, according
to Semantic Versioning, it's not necessary to increment the major version.