Refactored a test, according to the discussion at: https://github.com/GreanTech/AtomEventStore/pull/102/files#r19240650.
The goal of this particular process is to produce the Scan method. The Scan
method is the System Under Test.

The Scan method returns an instance of DataContractContentSerializer and
the behaviour of that class doesn't change as a result of this test being
added.

To make it explicit, the System Under Test is the Scan method. That's the
method added after having written the test. There's no sut object, because
the SUT isn't an object, but a static method.

By convention, we use the variable name actual for the result of exercising
the SUT. Once we have an actual variable, we should be done with
the Act phase.

The rest is verification.