Suppressed exceptions when updating the index.
The motivation is that when a new page is created, the second of two
update operations may fail, which means that the index document's last
link becomes stale. The system is still 'almost consistent' in the sense
that all other links point to correct documents, apart from the last link
in the index document. This link points to an existing document, but that
document is no longer the most recent document.

However, if the update operation of the index throws an exception, a
client may interpret that as a failure to write an event to storage, and
thus may retry the operation. This would be an error, because the event
has been written, and thus, a duplicate event would be written if the
operation is retried.

For that reason, while the first update operation (of the new previous
page) must succeed or throw an exception, the second update (of the index
page) must silently fail if an error occurs.